test(sage): run the Code Review Sage suite on Windows too
The suite was collected nowhere on Windows: conftest.py disabled
collection for the whole directory. The gate predates the app's own
Windows support work and conflated two things -- the app-level refusal
in discovery.py's gh_bin (tracked separately) and three harness details
that the tests themselves can express.

Lift the collection gate and give every platform-dependent test the
guard it actually needs:

- the planted-symlink tests now share one SYMLINKS_OK probe (moved from
  test_followup.py into tests/fixtures.py, the suite's shared module)
  and skip only where creating a symlink needs a privilege the host
  does not grant; the guards they pin keep running everywhere else;
- the two owner-only mode-bit assertions that are unobservable through
  st_mode on Windows keep their POSIX-only gates, and the report-output
  privacy test now pins its mode bits on POSIX while still asserting
  file presence and temp-file cleanup on every platform.

Measured on Windows 11: before, 21 failed / 735 passed / 11 skipped
with the gate lifted; after, 736 passed / 31 skipped / 0 failed, in
both serial and -n4 runs.
