+++ src/kiro_crew/apps/builtins/code_review_sage/tests/conftest.py
@@ -1,31 +1,12 @@
-"""Collection-time platform gate for the Code Review Sage suite.
+"""Shared fixtures for the Code Review Sage suite.
 
-The app itself now runs on Windows: the provider-CLI trust gate it shares with
-Issue Radar answers from the Windows ACL, and the review worker is handed the
-absolute interpreter the app resolves rather than the bare `python3` that is not
-an interpreter there.
-
-What still gates the suite is the suite, not the app. These tests assert POSIX
-behaviour throughout — `0600` file modes as `st_mode` bits (Windows expresses
-owner-only as a DACL and always reports `0o666`), forward-slash path suffixes,
-and shell-script `gh` stubs the Windows runner cannot execute — so running them
-there would fail on the harness rather than on anything under test. Making them
-Windows-native is separate work from making the app run.
-
-This lives next to the suite it gates, so the reason travels with the tests
-rather than sitting in a CI workflow that would hide it.
+The suite collects on every platform. Tests that pin inherently POSIX
+behaviour carry their own ``skipUnless`` guards (owner-only mode bits,
+unprivileged symlinks), so this module carries no platform gate -- only the
+fixtures every platform needs.
 """
 
-import os
-
 import pytest
-
-collect_ignore_glob = ["*"] if os.name == "nt" else []
-
-pytestmark = pytest.mark.skipif(
-    os.name == "nt",
-    reason="Code Review Sage's test harness is POSIX-only (see this conftest's docstring)",
-)
 
 
 @pytest.fixture(autouse=True)
